+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - **ENH: Add a `vectors.write` function to automatically detect the driver from the filename (and add the KML driver if needed)**
 - **ENH: Add shapely 2.0 functions in `ci.assert_geom_xxx` to handle more cases (2D/3D vectors, geomtry written in another way)**
+- FIX: Use our own function for setting the nodata instead of using rioxarray's in `rasters.crop`
 - DEPS: Pin Shapely >= 2.0
 
 ## 1.25.0 (2023-04-04)